Branch and BranchInjector now in separate file using CommonJS because previous change broke Mocha test environment
