+++ app/javascript/src/component_branch.js
@@ -7,13 +7,12 @@
  *
  * Documentation:
  *
- *     - jQueryUI
- *       https://api.jqueryui.com/dialog
- *
  *     - TODO:
+ *       /documentation/services/editor/javascript/component-branch.html
  *       (steven.burnell@digital.justice.gov.uk to add).
  *
  **/
+
 
 const utilities = require('./utilities');
 const BranchDestination = require('./component_branch_destination');
@@ -454,37 +453,6 @@ class BranchRemover {
 }
 
 
-/* Creates a BranchInjector object from passed $node.
- * BranchInjectors fetch new HTML for a branch that will
- * be added to the DOM and turned into a Branch object.
- **/
-class BranchInjector {
-  #config;
-
-  constructor($node, config) {
-    var injector = this;
-    var conf = utilities.mergeObjects({}, config);
-
-    this.#config = conf;
-    $node.on("click", function(e) {
-      e.preventDefault();
-      injector.add();
-    });
-  }
-
-  add() {
-    // This should really be something like `new Branch(); and stuff` here
-    // but, because much of it is haneled by the view and design requirements,
-    // we're just triggering an event to allow the document to know it is
-    // required. The view controller can pick up on this request for action
-    // and make DOM adjustments based on what it needs to happen, passing
-    // appropriate params to the available `new Branch()` fucntionality.
-    $(document.body).trigger("BranchInjector_Add");
-  }
-}
-
-
-
 // Make available for importing.
-export { Branch, BranchInjector }
+module.exports = Branch;
